requirements.txt
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==9.5.0
streamlit==1.26.0
tensorflow==2.13.0
filesplit==4.0.1